+++ TagCloudContainer/Infrastructure/Common/IAppSettings.cs
@@ -1,3 +1,3 @@
-﻿namespace TagCloudContainer.Infrastructure.Common;
+﻿namespace TagCloud.Infrastructure.Common;
 
 public interface IAppSettings : IImageSettingsProvider, IInputPathProvider, IOutputPathProvider, IFontProvider { }
