+++ TagCloudContainer/Infrastructure/Filter/IFilter.cs
@@ -1,13 +1,13 @@
-﻿using TagCloudContainer.Infrastructure.Lemmatizer;
+﻿using TagCloud.Infrastructure.Lemmatizer;
 
-namespace TagCloudContainer.Infrastructure.Filter;
+namespace TagCloud.Infrastructure.Filter;
 
 public interface IFilter
 {
     IEnumerable<string> FilterWords(IEnumerable<string> words);
     IEnumerable<string> FilterWords(IEnumerable<Lemma> lemmas);
 
-    Filter AddCondition(Func<string, bool> filter);
+    TagCloud.Infrastructure.Filter.Filter AddCondition(Func<string, bool> filter);
 
-    Filter AddCondition(Func<Lemma, bool> filter);
+    TagCloud.Infrastructure.Filter.Filter AddCondition(Func<Lemma, bool> filter);
 }
