+++ TagCloudContainer/Infrastructure/Layouter/ICloudLayouter.cs
@@ -1,6 +1,6 @@
 ﻿using System.Drawing;
 
-namespace TagCloudContainer.Infrastructure.Layouter;
+namespace TagCloud.Infrastructure.Layouter;
 
 public interface ICloudLayouter
 {
