+++ TagCloudContainer/Infrastructure/Lemmatizer/ILemmatizer.cs
@@ -1,4 +1,4 @@
-﻿namespace TagCloudContainer.Infrastructure.Lemmatizer;
+﻿namespace TagCloud.Infrastructure.Lemmatizer;
 
 public interface ILemmatizer
 {
