+++ TagCloudContainer/Program.cs
@@ -1,16 +1,16 @@
 ﻿using Autofac;
 using CommandLine;
-using TagCloudContainer.App.UI;
-using TagCloudContainer.Infrastructure.Common;
-using TagCloudContainer.Infrastructure.FileReader;
-using TagCloudContainer.Infrastructure.Filter;
-using TagCloudContainer.Infrastructure.Layouter;
-using TagCloudContainer.Infrastructure.Lemmatizer;
-using TagCloudContainer.Infrastructure.Painter;
-using TagCloudContainer.Infrastructure.Saver;
-using TagCloudContainer.Infrastructure.WordWeigher;
+using TagCloud.App.UI;
+using TagCloud.Infrastructure.Common;
+using TagCloud.Infrastructure.FileReader;
+using TagCloud.Infrastructure.Filter;
+using TagCloud.Infrastructure.Layouter;
+using TagCloud.Infrastructure.Lemmatizer;
+using TagCloud.Infrastructure.Painter;
+using TagCloud.Infrastructure.Saver;
+using TagCloud.Infrastructure.WordWeigher;
 
-namespace TagCloudContainer;
+namespace TagCloud;
 
 public static class Program
 {
