+++ TagCloudContainerTests/WordWeigherTestCases.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
 
-namespace TagCloudContainerTests;
+namespace TagCloudTests;
 
 internal class WordWeigherTestCases
 {
